fixed an incompatibility with stdint

+++ source/fundamental_types.h
@@ -13,18 +13,18 @@
 		#define inline __forceinline
 	#endif
 
-	typedef char int8_t;
+	typedef signed char int8_t;
 	typedef unsigned char uint8_t;
 
-	typedef short int16_t;
+	typedef signed short int16_t;
 	typedef unsigned short uint16_t;
 
-	typedef long int32_t;
-	typedef unsigned long uint32_t;
+	typedef int int32_t;
+	typedef unsigned int uint32_t;
 
 	#define UINT32_MAX  0xffffffff
 
-	typedef long long int64_t;
+	typedef signed long long int64_t;
 	typedef unsigned long long uint64_t;
 #else
 	#include <stdint.h>
